workflows: Calculate image filenames conditionally to be more like Noble (#38)
Imaage naming changed between Jammy and Noble when we moved to Github
actions for building, but provisioning expects images in oem-share to be named
according to a regex that matches the Jammy version.

This commit brings the naming back to look more like the Jammy version
and match the regex, while keeping the important changes that we've
made; notably:

- Not including the hardware name (g1200) in the UFS image names.
- Not including the internal "baoshan" project name in the image names.

Importantly, the "classic" moniker has been inserted conditionally for
server and desktop builds since the regex expects it. We did not need to
rename the boot assets in this commit since they continue to match the
regex that provisioning uses.

Closes: #36