workflows: Fix OEM Share upload breakage caused by 32d192b [issue 38] (#40)
The name of images changed but the wildcard to the artifact download had
not been updated. This meant that the images were not fetched before
the oem-share upload.

This commit addresses the problem by making the image name step an
action and invoking it in the build and upload jobs. It couldn't just be
a variable since both jobs use matrix inputs and the image name needs to
be re-calculated each time.

Closes: #39