workflows: Package images in a directory like provisioning expects (#48)
provisioning's deploy.py file calculates a directory name from the tarball
name, but prunes the first period and everything after it. It expects
that the *.img files will be in that directory after unpacking, an it
attempts to unpack the boot artifacts into the same directory.

Currently, provisioning cannot provision the Noble images because the tarball
does not look like this.

This commit just creates the directory in the tarball and gets on with
things.

Closes: PEB-1224