classic: use upstream seeds in noble image
Because the blocking from SRU team, we are unable to upload all required
packages to the Ubuntu archive before the noble GA release.
We are also unable to merge our Ubuntu seeds into the upstream. Because
not all pacakges are uploaded to Ubuntu archive.

Therefore, revert to use upstream seed.

This patch replace the seed from our downstream seeds to upstream seeds.
And all the preinstalled packages are added into extra-packages.

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>