workflows: add daily build for generic devel images (#80)
This commit aligns the generic image build with the existing daily build
workflow. The images will now be built and uploaded to oem-share daily
for testing on Genio EVKs.

The artifact names in all workflows are aligned with the daily image
convention to ensure they are found by the upload job.