classic: replace oem-baoshan-genio-meta with mediatek-genio-defaults
This patch replaces oem-baoshan-genio-desktop-meta and oem-baoshan-genio-meta
with mediatek-genio-defaults.

mediatek-genio-defaults is a new package on noble, and it will replace those
two "meta" packages.

This new package will merge desktop meta package and server meta package into
one. And we will maintain only one packages for both desktop image and server
image. So when user install 'ubuntu-desktop' packages on their server image,
they will not miss any workaround we applied.

Furthermore, this package is designed to be picked up by ubuntu-drivers (by
adding MODALIASES to control file), which will enable the generic ubuntu
installer to automatically install it in the future.

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>