workflows: Add a new tf-provision workflow to try booting an image (#60)
A new find-boot-image action will look in a directory and locate the
correct image and boot assets for the given hardware. The
tel-image-cache urls are returned for them.

The new tf-provision action will use the disk image and correct hardware
boot assets to provision a device via testflinger.

It will then fling a job to the queue requested and wait for it to
install, log in and download some logs to prove that it came up.

The new workflow is hooked into the build workflow so the daily images
are built as soon as they're uploaded. The workflow could accept
parameters, but for now it just provisions the daily build.

Machines are selected from the pool that Partner Engineering has
dedicated to this purpose. The G350 has a USB-Ethernet adaptor
connected, and so we do not need to set its onboard MAC address in
the workflow.

We can extend this to run some smoke tests later.

Important notes:

- Server is installed currently. Core should work but isn't tested and
  desktop appeared to boot to a login loop (that is ZAP-1277).  We can
  add core + desktop at a later date easily enough.

Closes: PEB-1118