workflows: separate boot assets build from daily build workflow (#122)
Separate the boot assets build from our daily build workflow and make it weekly.

Weekly boot assets would be uploaded to different dir to avoid overwriting the checksum file.

For release purpose, the original build logics are copied to `build-release-deploy.yml`.
It can build and upload the Ubuntu images & boot assets to the same directory.

The new workflows' badges are added to the README head.

Closes: PEB-1483