workflows: filter boot assets type before uploading (#125)
Currently, all boot assets are downloaded and uploaded to the same directory.
When uploading both images and boot assets to a shared location (for release),
this includes boot assets for image types that are not required.

This commit fixes this issue by filtering files based on their type
(core vs. non-core) before uploading.

Closes: PEB-1487